Remove and move files for navigation
