+++ src/AppRoutes.tsx
@@ -3,22 +3,22 @@ import { Route, Routes } from "react-router-dom";
 import App from "./views/App/App";
 import About from "./views/About/About";
 import MyProfile from "./views/MyProfile/MyProfile";
-import PhotoUpload from "./views/Intern/PhotoUpload/PhotoUpload";
-import Arkivsjef from "./views/Intern/Arkivsjef/ArchiveBoss";
+import PhotoUpload from "./views/Fg/PhotoUpload/PhotoUpload";
+import Arkivsjef from "./views/Fg/Arkivsjef/ArchiveBoss";
 import NotFound from "./views/NotFound/NotFound";
 import Search from "./views/Search/Search";
 import CsaTester from "./views/CsaTester";
-import Motives from "./views/Intern/Motives/Motives";
+import Motives from "./views/Fg/Motives/Motives";
 import Login from "./views/Login/AzureLogin";
-import EditMotive from "./views/Intern/EditMotive/EditMotive";
-import InternNav from "./views/Intern/InternNav/InternNav";
-import Expo from "./views/Intern/Expo/Expo";
+import EditMotive from "./views/Fg/EditMotive/EditMotive";
+import InternNav from "./views/Fg/FgNav/InternNav";
+import Expo from "./views/Fg/Expo/Expo";
 import Redirect from "./utils/Redirect/Redirect";
 import MotiveHeader from "./components/ImageViewer/MotiveHeader";
-import InternSearchView from "./views/Intern/InternSearch/InternSearchView";
+import InternSearchView from "./views/InternSearch/InternSearchView";
 import { Box } from "@mui/material";
 import { AuthenticationContext } from "./contexts/AuthenticationContext";
-import ArchiveBossEditUser from "./views/Intern/ArchiveBossEditUser/ArchiveBossEditUser";
+import ArchiveBossEditUser from "./views/Fg/ArchiveBossEditUser/ArchiveBossEditUser";
 import DeNyeSiden from "./views/DeNyeSiden/DeNyeSiden";
 import Photos from "./views/Photos/Photos";
 
@@ -38,12 +38,12 @@ const AppRoutes: FC = () => {
         <Route path="/logg-inn" element={<Login />} />
         {isAuthenticated && (
           <>
-            <Route path="/fg" element={<InternNav />} />
-            <Route path="/fg/myprofile" element={<MyProfile />} />
-            <Route path="/fg/search" element={<InternSearchView />} />
+            <Route path="/intern/search" element={<InternSearchView />} />
             {position === "FG" && (
               <>
-                <Route path="/fg/DeNyeSiden" element={<DeNyeSiden />} />
+                <Route path="/fg" element={<InternNav />} />
+                <Route path="/fg/myprofile" element={<MyProfile />} />
+                <Route path="/fg/projects" element={<DeNyeSiden />} />
                 <Route path="/fg/last-opp" element={<PhotoUpload />} />
                 <Route path="/fg/arkivsjef" element={<Arkivsjef />} />
                 <Route
@@ -65,16 +65,6 @@ const AppRoutes: FC = () => {
         <Route
           path="/fg-wiki"
           element={<Redirect link="https://wiki.samfundet.no/fg/" />}
-        />
-        <Route
-          path="/utlaan"
-          element={
-            <Redirect link="https://wiki.samfundet.no/fg/Utl%C3%A5nsordning" />
-          }
-        />
-        <Route
-          path="/fg-web"
-          element={<Redirect link="https://github.com/Fotogjengen/" />}
         />
       </Routes>
     </Box>
