+++ src/views/Fg/Expo/PhotoGridItem.tsx
