+++ src/views/InternSearch/InternSearchView.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import styles from "./InternSearch.module.css";
 import InternSearchInput from "./InternSearchInput";
 import CustomDataGrid from "./CustomTable";
-import { PhotoApi, PhotoSearch } from "../../../utils/api/PhotoApi";
-import { PhotoDto } from "../../../../generated";
+import { PhotoApi, PhotoSearch } from "../../utils/api/PhotoApi";
+import { PhotoDto } from "../../../generated";
 
 
 const InternSearchView = () => {
